Add login and register pages with Firebase auth
